Updated response models.
Map and Game Variants properties now appear on the BaseMatch model
(previously only on MatchSet). PlayerScore added to all Match types
models (appears to only ever set to null).
