+++ Source/HaloSharp.Test/JSON/Stats/Common/variant.schema.json
@@ -1,0 +1,24 @@
+{
+  "type": [ "object", "null" ],
+  "additionalProperties": false,
+  "properties": {
+    "ResourceType": {
+      "type": "integer"
+    },
+    "ResourceId": {
+      "type": "string"
+    },
+    "OwnerType": {
+      "type": "integer"
+    },
+    "Owner": {
+      "type": "string"
+    }
+  },
+  "required": [
+    "ResourceType",
+    "ResourceId",
+    "OwnerType",
+    "Owner"
+  ]
+}
