+++ Source/HaloSharp/HaloSharp.csproj
@@ -113,6 +113,7 @@
     <Compile Include="Model\Stats\Lifetime\ArenaServiceRecord.cs" />
     <Compile Include="Model\Stats\Common\Link.cs" />
     <Compile Include="Model\Stats\MatchSet.cs" />
+    <Compile Include="Model\Stats\Common\Variant.cs" />
     <Compile Include="Model\ValidationResult.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Query\Metadata\GetSeasons.cs" />
